requirements.txt
huggingface-hub==0.27.0
python-dotenv==1.0.1
streamlit==1.41.1

